Compile string partials, make all compiled templates available as partials
Previously `renderStringTemplate` passed `this.compiledTemplates` as partials
and `renderNamedTemplate` passed `data.partials`. The latter was failing with
strings ("No compiler available") and didn't work passing compiled templates as
could not be serialized as event data.

Now all string partials are compiled and the `this.compiledtemplates`
dictionary is always passed as partials to `render`.

Added tests and brief explanation to README. Also upgraded karma to 0.10.x as
couldn't get tests to run on 0.8.x.